put the new function in Utils.java
