+++ gazeplay-commons/src/main/java/net/gazeplay/commons/utils/games/Utils.java
@@ -1,5 +1,8 @@
 package net.gazeplay.commons.utils.games;
 
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
+import javafx.scene.control.TextField;
 import lombok.extern.slf4j.Slf4j;
 import net.gazeplay.commons.configuration.ActiveConfigurationContext;
 import net.gazeplay.commons.configuration.Configuration;
@@ -75,5 +78,22 @@ public class Utils {
             final String element = (String) propertyNames.nextElement();
             log.info(String.format("%s: %s", element, System.getProperty(element)));
         }
+    }
+
+    /**
+     *
+     * @param tf textfield to limit the number of characters
+     * @param maxLength the max number of characters allowed
+     */
+    public static void addTextLimiter(final TextField tf, final int maxLength) {
+        tf.textProperty().addListener(new ChangeListener<String>() {
+            @Override
+            public void changed(final ObservableValue<? extends String> ov, final String oldValue, final String newValue) {
+                if (tf.getText().length() > maxLength) {
+                    String s = tf.getText().substring(0, maxLength);
+                    tf.setText(s);
+                }
+            }
+        });
     }
 }
